refactor: 미사용 클래스 제거 및 패키지 구조 변경
